backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
langchain==0.0.310
langchain-community==0.0.6
langchain-google-genai==0.0.5
google-generativeai==0.3.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
flask-caching==2.1.0
redis==5.0.1
